refactor(runs): split RunDetailPanel — 4 tab bodies to their own files (#1509)
Follow-up to #1507 + #1508. The RunDetailPanel is now the shared surface
for both the canvas run detail page and Lens RunBubble — a bug in one
tab affects both. Splitting each tab into its own file keeps reviews
scoped and reduces the odds of unrelated regressions.

Files extracted (no behavior change, same props they had inline):
- apps/web/src/components/runs/tabs/RunSummaryTab.tsx   (62 lines)
- apps/web/src/components/runs/tabs/RunFilesTab.tsx     (82 lines)
- apps/web/src/components/runs/tabs/RunApprovalsTab.tsx (126 lines)
- apps/web/src/components/runs/tabs/RunCostTab.tsx      (74 lines)

RunDetailPanel.tsx: 636 -> 418 lines (state, effects, header, StatRow,
tab shell, action buttons). Steps + AI Trace tabs already delegate to
RunTrace / ConversationTrace and are unchanged.

The RunApprovalsTab body was itself decomposed into three internal
subcomponents (AwaitingReview, DecisionRecorded, PriorDecisions) — the
original nested-ternary was hard to reason about. Same rendering paths.

RunMeta is now exported from RunDetailPanel so tab files can type their
`run` prop. Idempotent with the same export in the open #1508 follow-up.

Verify:
- tsc --noEmit: clean
- eslint: 0 errors on all changed files; only the pre-existing
  react-hooks/exhaustive-deps warning on RunDetailPanel (moved from
  line 134 -> 131 since the file shrank).
